Get a default scale for correlation heatmaps.

Get a default scale for correlation heatmaps.

Usage

default_scale_corr(
  aes_type,
  bins = NULL,
  limits = c(-1, 1),
  size_range = NULL,
  high = "sienna2",
  mid = "white",
  low = "skyblue2",
  midpoint = 0,
  na_col = "grey50",
  leg_order = 1,
  title = ggplot2::waiver()
)

Arguments

aes_type

Type of aesthetic ('fill', 'col', or 'size').

bins

Number of bins (for fill and colour scales).

limits

Scale limits (fill and colour).

size_range

Range of sizes (size).

high

Colours at higher end of fill or colour scale.

mid

Colours at middle point of fill or colour scale.

low

Colours at lower end of fill or colour scale.

midpoint

Middle point of colour scale.

na_col

Colour of NAs.

leg_order

Order of legend.

title

Title of legend.

Value

ggplot2 scale for correlation heatmap.