Orient a dendrogram.

Orient a dendrogram.

Usage

orient_dendrogram(dend, dim = c("rows", "cols"), full_plt, layout, dend_side)

Arguments

dend

Dendrogram segments or nodes data frame (containing x, y, xend, yend).

dim

String, rows or cols to know which dimensions dendrogram should be on.

full_plt

Logical indicating if it's for the full layout.

layout

The heatmap layout to take mixed layout into account.

dend_side

Logical indicating if the dendrogram should be placed on the left (if row dend) or bottom (if col dend).

Value

The input dendrogram data frame but rotated and mirrored to fit the plot.