Prepare scales for heatmap.

Prepare scales for heatmap.

Usage

prepare_scales(
  scale_order,
  context = c("gghm", "ggcorrhm"),
  layout,
  val_type,
  col_scale = NULL,
  col_name = "value",
  size_scale = NULL,
  size_name = "value",
  bins = NULL,
  limits = c(-1, 1),
  high = "sienna2",
  mid = "white",
  low = "skyblue2",
  midpoint = 0,
  size_range = NULL,
  na_col = "grey50"
)

Arguments

scale_order

List of necessary scales and their orders, as obtained make_legend_order.

context

Scale context (gghm or ggcorrhm) for deciding which defaults to use.

layout

Layout of plot to treat parameters depending on length.

val_type

String with type of value ('continuous' or 'discrete').

col_scale

Colour scales input.

col_name

Colour scale names.

size_scale

Size scales input.

size_name

Size scale names.

bins

Number of bins to divide scale into (only for ggcorrhm).

limits

Scale limits (for ggcorrhm).

high

Colour at higher end of scale (for ggcorrhm).

mid

Colour at middle point of scale (for ggcorrhm).

low

Colour at Lower end of scale (for ggcorrhm).

midpoint

Middle point of scale (for ggcorrhm).

size_range

Size range for size scale (for ggcorrhm).

na_col

Colour to use for NAs (for ggcorrhm).

Value

ggplot2 scales for the plot.